+++ packages/tswebextension/test/lib/mv2/background/allowlist.test.ts
@@ -38,7 +38,7 @@ describe('Allowlist Api', () => {
                 inverted: false,
                 expected: new StringRuleList(
                     ALLOWLIST_FILTER_ID,
-                    '@@///(www\\.)?example.com/$document,important',
+                    '@@///(www\\.)?example\\.com/$document,important',
                 ),
             },
             {
@@ -80,11 +80,9 @@ describe('Allowlist Api', () => {
 
     describe('static createAllowlistRule method', () => {
         it('should return allowlist rule, when domain is specified', () => {
-            const domain = 'example.com';
-
-            expect(Allowlist.createAllowlistRule(domain)).toStrictEqual(
+            expect(Allowlist.createAllowlistRule('example.com')).toStrictEqual(
                 new NetworkRule(
-                    String.raw`@@///(www\.)?${domain}/$document,important`,
+                    String.raw`@@///(www\.)?example\.com/$document,important`,
                     ALLOWLIST_FILTER_ID,
                 ),
             );