+++ packages/tswebextension/src/lib/mv2/background/tabs/tabs-cosmetic-injector.ts
@@ -4,10 +4,7 @@ import { RequestType } from '@adguard/tsurlfilter/es/request-type';
 import { isHttpOrWsRequest } from '../../../common/utils/url';
 import { logger } from '../../../common/utils/logger';
 import { ContentType } from '../../../common/request-type';
-import {
-    type ApplyCssRulesParams,
-    CosmeticApi,
-} from '../cosmetic-api';
+import { CosmeticApi } from '../cosmetic-api';
 import { Frame, MAIN_FRAME_ID } from './frame';
 import { TabContext } from './tab-context';
 import type { EngineApi } from '../engine-api';
@@ -105,19 +102,9 @@ export class TabsCosmeticInjector {
 
             const { cosmeticResult } = frame;
 
-            const cssInjectionParams: ApplyCssRulesParams = {
-                tabId,
-                frameId,
-                cosmeticResult,
-            };
+            CosmeticApi.applyFrameCssRules(frameId, tabId);
 
-            CosmeticApi.applyFrameCssRules(cssInjectionParams);
-
-            CosmeticApi.applyFrameJsRules({
-                tabId,
-                frameId,
-                cosmeticResult,
-            });
+            CosmeticApi.applyFrameJsRules(frameId, tabId);
 
             CosmeticApi.logScriptRules({
                 url,