+++ packages/tswebextension/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: manually add compare links for version to the end of the file -->
 <!-- e.g. [0.1.2]: https://github.com/AdguardTeam/tsurlfilter/compare/tswebextension-v0.1.1...tswebextension-v0.1.2 -->
 
+## Unreleased
+
+### Fixed
+- Domains from the allowlist were not properly escaped before being passed into the regular expression rules. [2461](https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2461)
 
 ## [0.3.13] - 2023-08-25